scripts/requirements.txt
python-dotenv
pyjwt[crypto]
cryptography
